added autentication to access workout for only authenticated users , added route protection , created check for user logins

+++ frontend/src/components/WorkoutDetails.js
@@ -1,14 +1,23 @@
-import { useWorkoutsContext } from "../hooks/useWorkoutsContext" 
+import { useWorkoutsContext } from '../hooks/useWorkoutsContext'
+import { useAuthContext } from '../hooks/useAuthContext'
 
 // date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext()
+  const { user } = useAuthContext()
 
   const handleClick = async () => {
+    if (!user) {
+      return
+    }
+
     const response = await fetch('/api/workouts/' + workout._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
     })
     const json = await response.json()
 
@@ -18,12 +27,12 @@ const WorkoutDetails = ({ workout }) => {
   }
 
   return (
-    <div className='workout-details'>
-        <h4>{workout.title}</h4>
-        <p><strong>Load (kg): </strong>{workout.load}</p>
-        <p><strong>Reps: </strong>{workout.reps}</p>
-        <p>{formatDistanceToNow(new Date(workout.createdAt), {addSuffix: true})}</p>
-        <span className="material-symbols-outlined" onClick={handleClick}>Delete</span>
+    <div className="workout-details">
+      <h4>{workout.title}</h4>
+      <p><strong>Load (kg): </strong>{workout.load}</p>
+      <p><strong>Reps: </strong>{workout.reps}</p>
+      <p>{formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}</p>
+      <span className="material-symbols-outlined" onClick={handleClick}>delete</span>
     </div>
   )
 }
